+++ usuario/urls.py
@@ -7,6 +7,7 @@ urlpatterns = [
     path('cadastrar_usuario/', views.cadastrar_usuario, name='cadastrar_usuario'),
     path("login/", views.login, name="login"),
     path('editar-usuario/', views.editar_usuario, name='editar_usuario'),
+    path('usuarios/<int:usuario_id>/hosts/', views.hosts_usuario, name='hosts_usuario'),
     path('logout/', views.logout, name='logout'),
     path('trocar-senha/', views.trocar_senha, name='trocar_senha'),
     path('2fa/', views.configurar_2fa, name='configurar_2fa'),